Make CreateNotification an explicit interface on notifications.Service.
This changes a number of methods on vulnerabilities.Service to take a
notifications.Service instance to facilitate sending notifications.

This introduces a notifications.Null implementation of
notifications.Service that supports CreateNotification into the ether,
and returning an error for all other methods.  This is used in places
where vulnerabilities.Service methods already took a bool to suppress
notifications.