Make the notification service a bit thinner.
Previously GetNotification on notifications.Service would read
Vulnerability data and Layer data in order to populate the paginated
list of layers introducing a given vulnerability.

This change adjusts the GetNotification method to take fewer args and
return a skeletal Vulnerability (only Model populated) and place the
burden of fetching the full Vulnerability and paginated Layer list on
the caller.

This removes the need for notifications.Service to have a handle to
vulnerabilities.Service and layers.Service.

As part of this, layers.Service grew a new method (exposed from
notifications) for loading the paginated layer-introduced-in list into
a Vulnerability.